+++ reana-ui/src/pages/workflowDetails/components/WorkflowInfo.js
@@ -2,7 +2,7 @@
   -*- coding: utf-8 -*-
 
   This file is part of REANA.
-  Copyright (C) 2020 CERN.
+  Copyright (C) 2020, 2022 CERN.
 
   REANA is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
@@ -15,7 +15,7 @@ import { Icon, Popup } from "semantic-ui-react";
 import { INTERACTIVE_SESSION_URL } from "~/client";
 import { getReanaToken } from "~/selectors";
 import { statusMapping } from "~/util";
-import { WorkflowProgress } from "../components";
+import { LauncherLabel, WorkflowProgress } from "../components";
 import { JupyterNotebookIcon, WorkflowActionsPopup } from "~/components";
 import styles from "./WorkflowInfo.module.scss";
 
@@ -35,6 +35,7 @@ export default function WorkflowInfo({ workflow }) {
     completed,
     total,
     status,
+    launcherURL,
     session_uri: sessionUri,
     session_status: sessionStatus,
   } = workflow;
@@ -44,23 +45,28 @@ export default function WorkflowInfo({ workflow }) {
     <div className={styles.workflow}>
       <section className={styles.info}>
         <div>
-          <Icon
-            name={statusMapping[status].icon}
-            color={statusMapping[status].color}
-          />{" "}
-          <span className={styles["name"]}>{name}</span>
-          <span className={styles["run"]}>#{run}</span>
-          {isSessionOpen && (
-            <a
-              href={INTERACTIVE_SESSION_URL(sessionUri, reanaToken)}
-              target="_blank"
-              rel="noopener noreferrer"
-              onClick={(e) => e.stopPropagation()}
-              className={styles.notebook}
-            >
-              <JupyterNotebookIcon />
-            </a>
-          )}
+          <div className={styles["first-row"]}>
+            <Icon
+              name={statusMapping[status].icon}
+              color={statusMapping[status].color}
+            />{" "}
+            <span className={styles["name"]}>{name}</span>
+            <span className={styles["run"]}>#{run}</span>
+            {isSessionOpen && (
+              <a
+                href={INTERACTIVE_SESSION_URL(sessionUri, reanaToken)}
+                target="_blank"
+                rel="noopener noreferrer"
+                onClick={(e) => e.stopPropagation()}
+                className={styles.notebook}
+              >
+                <JupyterNotebookIcon />
+              </a>
+            )}
+            <span className={styles["launcher-label"]}>
+              <LauncherLabel url={launcherURL} />
+            </span>
+          </div>
           <Popup
             trigger={
               <div>
@@ -82,13 +88,15 @@ export default function WorkflowInfo({ workflow }) {
         </div>
         <div className={styles.info}>
           <div>
-            <span
-              className={`${styles["status"]} sui-${statusMapping[status].color}`}
-            >
-              {status}
-            </span>{" "}
-            {statusMapping[status].preposition}{" "}
-            {status !== "deleted" && duration}
+            <div className={styles["first-row"]}>
+              <span
+                className={`${styles["status"]} sui-${statusMapping[status].color}`}
+              >
+                {status}
+              </span>{" "}
+              {statusMapping[status].preposition}{" "}
+              {status !== "deleted" && duration}
+            </div>
             {NON_FINISHED_STATUSES.includes(status) && (
               <Icon
                 name="refresh"