+++ reana-ui/src/util.js
@@ -2,7 +2,7 @@
   -*- coding: utf-8 -*-
 
   This file is part of REANA.
-  Copyright (C) 2020 CERN.
+  Copyright (C) 2020, 2022 CERN.
 
   REANA is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
@@ -53,6 +53,7 @@ export function parseWorkflows(workflows) {
     const total = workflow.progress.total;
     workflow.completed = typeof progress === "object" ? progress.total : 0;
     workflow.total = total.total;
+    workflow.launcherURL = workflow.launcher_url;
     workflow = parseWorkflowDates(workflow);
 
     obj[workflow.id] = workflow;